requirements.txt
diskcache==4.0.0
Flask==1.1.1
lxml
pyoembed==0.1.2
readability-lxml==0.7.1
requests==2.22.0
